src/tools/azure-cli-extension/cleanroom/requirements.txt
azdev==0.1.67
azure-core==1.30.1
azure-cli-core==2.60.0
python-on-whales==0.71.0
pycryptodome==3.19.1
cryptography==44.0.1
docker
build==1.2.2-post1
oras==0.1.29